+++ projects/extension/src/containers/Popup.tsx
@@ -3,66 +3,31 @@ import { FunctionComponent, ReactNode, useEffect, useState } from "react"
 import { MdOutlineSettings, MdOutlineEast } from "react-icons/md"
 import { Accordion, Logo, IconWeb3, BraveModal } from "../components"
 import * as environment from "../environment"
+import { useActiveChains } from "../hooks/useActiveChains"
 
-interface PopupChain {
-  chainName: string
-  isWellKnown: boolean
-  details: ChainDetails[]
+const goToOptions = (): void => {
+  chrome.runtime.openOptionsPage()
 }
 
-interface ChainDetails {
-  tabId?: number
-  url?: string
-  peers: number
-  isSyncing: boolean
-  chainId: string
-  bestBlockHeight: number | undefined
+const NetworkIcon = ({
+  network,
+  isWellKnown,
+}: {
+  network: string
+  isWellKnown: boolean
+}) => {
+  const icon = network.toLowerCase()
+  return (
+    <>
+      <IconWeb3 isWellKnown={isWellKnown}>{icon}</IconWeb3>
+      <div className="pl-2">{network}</div>
+    </>
+  )
 }
 
 const Popup: FunctionComponent = () => {
-  const [connChains, setConnChains] = useState<PopupChain[] | undefined>()
+  const chains = useActiveChains()
   const [showModal, setShowModal] = useState<boolean>(false)
-
-  const refresh = () => {
-    environment.getAllActiveChains().then((chains) => {
-      const allChains: PopupChain[] = []
-      ;(chains || []).forEach((c) => {
-        const i = allChains.findIndex(
-          (i) => i.chainName === c.chainName && i.isWellKnown === c.isWellKnown,
-        )
-        const { peers, isSyncing, chainId, bestBlockHeight } = c
-        if (i === -1) {
-          allChains.push({
-            chainName: c.chainName,
-            isWellKnown: c.isWellKnown,
-            details: [
-              {
-                tabId: c.tab.id,
-                url: c.tab.url,
-                peers,
-                isSyncing,
-                chainId,
-                bestBlockHeight,
-              },
-            ],
-          })
-        } else {
-          const details = allChains[i]?.details
-          if (!details.map((d) => d.tabId).includes(c.tab.id)) {
-            details.push({
-              tabId: c.tab.id,
-              url: c.tab.url,
-              peers,
-              isSyncing,
-              chainId,
-              bestBlockHeight,
-            })
-          }
-        }
-      })
-      setConnChains([...allChains])
-    })
-  }
 
   useEffect(() => {
     // Identify Brave browser and show Popup
@@ -70,25 +35,7 @@ const Popup: FunctionComponent = () => {
       const braveSetting = await environment.get({ type: "braveSetting" })
       setShowModal(isBrave && !braveSetting)
     })
-
-    const unregister = environment.onActiveChainsChanged(() => refresh())
-    refresh()
-    return unregister
   }, [])
-
-  const goToOptions = (): void => {
-    chrome.runtime.openOptionsPage()
-  }
-
-  const networkIcon = (network: string, isWellKnown: boolean) => {
-    const icon = network.toLowerCase()
-    return (
-      <>
-        <IconWeb3 isWellKnown={isWellKnown}>{icon}</IconWeb3>
-        <div className="pl-2">{network}</div>
-      </>
-    )
-  }
 
   return (
     <>
@@ -101,29 +48,40 @@ const Popup: FunctionComponent = () => {
               Go to Options
             </span>
             <MdOutlineSettings
+              data-testid={"btnGoToOptions"}
               onClick={goToOptions}
               className="text-xl leading-5 cursor-pointer hover:bg-gray-200"
             />
           </div>
         </header>
-        <div className={!connChains?.length ? "" : "pb-3.5"}>
-          {!connChains?.length ? (
+        <div className={!chains?.length ? "" : "pb-3.5"}>
+          {!chains?.length ? (
             <div className="mx-8 my-8">
               The extension isn't connected to any network.
             </div>
           ) : (
-            connChains?.map((w) => {
-              if (w?.details?.length === 1 && !w?.details[0].tabId)
+            chains?.map((c) => {
+              if (c?.details?.length === 1 && !c?.details[0].tabId)
                 return (
                   <>
-                    <div className="block mt-4">
-                      <div key={w.chainName} className="pl-6 flex text-lg">
-                        {networkIcon(w.chainName, w.isWellKnown)}
+                    <div
+                      className="block mt-4"
+                      data-testid={`chain${c.chainName}`}
+                    >
+                      <div key={c.chainName} className="pl-6 flex text-lg">
+                        <NetworkIcon
+                          network={c.chainName}
+                          isWellKnown={c.isWellKnown}
+                        />
                       </div>
                       <div className="pl-[4.5rem] text-sm flex pt-2">
                         <span className="text-[#323232]">Latest block</span>
-                        <span className="pl-2 text-[#24CC85]">
-                          {w?.details[0].bestBlockHeight?.toLocaleString(
+                        <span
+                          className="pl-2 text-[#24CC85]"
+                          data-testid="blockheight"
+                          data-blockheight={c?.details[0].bestBlockHeight}
+                        >
+                          {c?.details[0].bestBlockHeight?.toLocaleString(
                             "en-US",
                           ) || "Syncing..."}
                         </span>
@@ -135,7 +93,7 @@ const Popup: FunctionComponent = () => {
                   </>
                 )
               const contents: ReactNode[] = []
-              w?.details?.forEach((t) => {
+              c?.details?.forEach((t) => {
                 if (t.tabId) {
                   contents.push(
                     <div key={t.url} className="flex justify-between">
@@ -153,17 +111,27 @@ const Popup: FunctionComponent = () => {
                   titleClass="popup-accordion-title"
                   contentClass="popup-accordion-content"
                   titles={[
-                    <div className="block mt-4">
+                    <div
+                      className="block mt-4"
+                      data-testid={`chain${c.chainName}`}
+                    >
                       <div className="pl-4 flex text-lg justify-start">
-                        {networkIcon(w.chainName, w.isWellKnown)}
+                        <NetworkIcon
+                          network={c.chainName}
+                          isWellKnown={c.isWellKnown}
+                        />
                         <span className="pl-2 text-[#616161]">
                           ({contents.length})
                         </span>
                       </div>
                       <div className="pl-16 flex pt-2">
                         <span className="text-[#323232]">Latest block</span>
-                        <span className="pl-2 text-[#24CC85]">
-                          {w?.details[0].bestBlockHeight?.toLocaleString(
+                        <span
+                          className="pl-2 text-[#24CC85]"
+                          data-testid="blockheight"
+                          data-blockheight={c?.details[0].bestBlockHeight}
+                        >
+                          {c?.details[0].bestBlockHeight?.toLocaleString(
                             "en-US",
                           ) || "Syncing..."}
                         </span>