+++ projects/extension/src/content/index.ts
@@ -1,31 +1,7 @@
-import {
-  DOM_ELEMENT_ID,
-  ToApplication,
-} from "@substrate/connect-extension-protocol"
+import { register } from "@polkadot-api/light-client-extension-helpers/content-script"
+import { DOM_ELEMENT_ID } from "@substrate/connect-extension-protocol"
 
-import checkMessage from "./checkMessage"
-import { PORTS } from "../shared"
-import type { ToBackground, ToContent } from "../protocol"
-
-// Set up a promise for when the page is activated,
-// which is needed for prerendered pages.
-const whenActivated = new Promise<void>((resolve) => {
-  // @ts-ignore
-  if (document.prerendering) {
-    // @ts-ignore
-    document.addEventListener("prerenderingchange", resolve)
-  } else {
-    resolve()
-  }
-})
-
-const portPostMessage = (
-  port: chrome.runtime.Port | undefined,
-  msg: ToBackground,
-) => port?.postMessage(msg)
-
-const windowPostMessage = (msg: ToApplication, origin: string) =>
-  window.postMessage(msg, origin)
+register(DOM_ELEMENT_ID)
 
 // inject as soon as possible the DOM element necessary for web pages to know that the extension
 // is available
@@ -40,81 +16,6 @@ window.document.addEventListener("readystatechange", () => {
   const s = document.createElement("span")
   s.id = DOM_ELEMENT_ID
   s.setAttribute("style", "display:none")
+  s.setAttribute("channelid", DOM_ELEMENT_ID)
   document.body.appendChild(s)
-
-  let port: chrome.runtime.Port | undefined
-
-  const chainIds = new Set<string>()
-  const handleExtensionError = (errorMessage: string, origin: string) => {
-    console.error(errorMessage)
-    chainIds.forEach((chainId) =>
-      windowPostMessage(
-        {
-          origin: "substrate-connect-extension",
-          chainId,
-          type: "error",
-          errorMessage,
-        },
-        origin,
-      ),
-    )
-    chainIds.clear()
-  }
-
-  window.addEventListener("message", async ({ data, source, origin }) => {
-    if (source !== window) return
-    if (data?.origin !== "substrate-connect-client") return
-    if (!checkMessage(data)) {
-      // probably someone abusing the extension
-      console.warn("Malformed message - unrecognised message.type", data)
-      return
-    }
-
-    await whenActivated
-
-    if (!port) {
-      try {
-        port = chrome.runtime.connect({ name: PORTS.CONTENT })
-      } catch (error) {
-        handleExtensionError(
-          "Cannot connect to substrate-connect extension",
-          origin,
-        )
-        return
-      }
-      port.onMessage.addListener((msg: ToContent) => {
-        if (msg.type === "keep-alive-ack") return
-        if (msg.type === "error") chainIds.delete(msg.chainId)
-        windowPostMessage(msg, origin)
-      })
-      const keepAliveInterval = setInterval(
-        () => portPostMessage(port, { type: "keep-alive" }),
-        20_000,
-      )
-      port.onDisconnect.addListener(() => {
-        port = undefined
-        clearInterval(keepAliveInterval)
-        handleExtensionError(
-          "Disconnected from substrate-connect extension",
-          origin,
-        )
-      })
-    }
-
-    portPostMessage(port, data)
-
-    switch (data.type) {
-      case "add-chain":
-      case "add-well-known-chain": {
-        chainIds.add(data.chainId)
-        break
-      }
-      case "remove-chain": {
-        chainIds.delete(data.chainId)
-        break
-      }
-      default:
-        break
-    }
-  })
 })