+++ projects/extension/src/environment.ts
@@ -1,40 +1,7 @@
-export type StorageEntry =
-  | { type: "braveSetting" }
-  | { type: "database"; chainName: string }
-  | { type: "bootnodes"; chainName: string }
-  | { type: "activeChains"; tabId: number }
+export type StorageEntry = { type: "braveSetting" }
 
 export type StorageEntryType<E extends StorageEntry> =
-  E["type"] extends "braveSetting"
-    ? boolean
-    : E["type"] extends "database"
-    ? string
-    : E["type"] extends "bootnodes"
-    ? string[]
-    : E["type"] extends "activeChains"
-    ? ExposedChainConnection[]
-    : never
-
-const activeChainsStorage = chrome.storage[storeageAreaOf("activeChains")]
-
-/**
- * Finds all the `activeChains` entries, and concatenates them together.
- *
- * Important note: the reason why each tab has its own local storage entry is to avoid race
- * conditions where tabs send messages at the same time and overwrite each other's list of chains.
- */
-export async function getAllActiveChains(): Promise<ExposedChainConnection[]> {
-  return new Promise((resolve) => {
-    activeChainsStorage.get(null, (res) => {
-      let out: ExposedChainConnection[] = []
-      for (const key in res) {
-        if (key.startsWith("activeChains_"))
-          out = [...out, ...(res[key] as ExposedChainConnection[])]
-      }
-      resolve(out)
-    })
-  })
-}
+  E["type"] extends "braveSetting" ? boolean : never
 
 export async function getDefaultBootnodes(
   chain: string,
@@ -50,20 +17,14 @@ export async function getDefaultBootnodes(
   return undefined
 }
 
-// Load default Bootnodes and save them to localStorage
-export async function getBootnodes(chainName: string) {
-  return await get({ type: "bootnodes", chainName })
-}
-
 export async function get<E extends StorageEntry>(
   entry: E,
 ): Promise<StorageEntryType<E> | undefined> {
   return new Promise((resolve) => {
     const key = keyOf(entry)
-    const area = storeageAreaOf(entry.type)
     // Note that `res[key]` will contain `undefined` is there is no such item in the
     // storage (tested on Chrome v106).
-    chrome.storage[area].get([key], (res) => resolve(res[key]))
+    chrome.storage.local.get([key], (res) => resolve(res[key]))
   })
 }
 
@@ -73,72 +34,13 @@ export async function set<E extends StorageEntry>(
 ): Promise<void> {
   return new Promise((resolve) => {
     const key = keyOf(entry)
-    const area = storeageAreaOf(entry.type)
-    chrome.storage[area].set({ [key]: value }, () => resolve())
+    chrome.storage.local.set({ [key]: value }, () => resolve())
   })
-}
-
-export async function remove<E extends StorageEntry>(entry: E): Promise<void> {
-  return new Promise((resolve) => {
-    const key = keyOf(entry)
-    const area = storeageAreaOf(entry.type)
-    chrome.storage[area].remove(key, () => resolve())
-  })
-}
-
-export async function clearAllActiveChains(): Promise<void> {
-  return new Promise((resolve) => {
-    activeChainsStorage.get(null, (res) => {
-      const keys = []
-      for (const key in res) {
-        if (key.startsWith("activeChains_")) keys.push(key)
-      }
-      activeChainsStorage.remove(keys, () => resolve())
-    })
-  })
-}
-
-export function onActiveChainsChanged(callback: () => void): () => void {
-  const registeredCallback = (changes: {
-    [key: string]: chrome.storage.StorageChange
-  }) => {
-    for (const key in changes) {
-      if (key.startsWith("activeChains_")) callback()
-    }
-  }
-  chrome.storage.onChanged.addListener(registeredCallback)
-  return () => chrome.storage.onChanged.removeListener(registeredCallback)
 }
 
 function keyOf(entry: StorageEntry): string {
   switch (entry.type) {
     case "braveSetting":
       return "braveSetting"
-    case "database":
-      return entry.chainName // TODO: change this to add a prefix
-    case "bootnodes":
-      return "bootNodes_" + entry.chainName
-    case "activeChains":
-      return "activeChains_" + entry.tabId
   }
-}
-
-function storeageAreaOf(type: StorageEntry["type"]): "local" | "session" {
-  // chrome.storage.session is available from Chrome v102+ with MV3+ and Firefox 115+
-  return chrome.storage.session && type === "activeChains" ? "session" : "local"
-}
-
-export interface ExposedChainConnection {
-  chainId: string
-  chainName: string
-  isWellKnown: boolean
-  tab: ExposedChainConnectionTabInfo
-  isSyncing: boolean
-  peers: number
-  bestBlockHeight?: number
-}
-
-export interface ExposedChainConnectionTabInfo {
-  id: number
-  url: string
 }